Move problem report fragment to fragment package
